+++ resources-ktx/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - **Context**: added type checks to extensions `resolveColor`
 - **Context**: new extensions `resolveDimension`, `resolveDimensionPixelSize` and `resolveDimensionPixelOffset`
-- **Context**: new extension `resolveBoolean`
+- **Context**: new extensions `resolveBoolean`, `resolveInt` and `resolveFloat`
 
 ### Dependencies